refactored test_threads
